requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.5
langchain==0.1.0
scikit-learn==1.3.2
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pytest==7.4.3
black==23.12.0
